:truck: rename: 디렉토리 구조 수정 및 패키지 이동
