+++ backend/src/main/java/cotato/backend/domains/post/dto/PostDto.java
@@ -1,0 +1,4 @@
+package cotato.backend.domains.post.dto;
+
+public record PostDto(Long id, String title, String content, String name, int views) {
+}
