+++ backend/src/main/java/cotato/backend/domains/repository/PostRepository.java
@@ -1,0 +1,10 @@
+package cotato.backend.domains.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import cotato.backend.domains.post.entity.Post;
+
+@Repository
+public interface PostRepository extends JpaRepository<Post, Long> {
+}
